ai-service/requirements.txt
# requirements.txt (Python 3.12 compatible)
flask==3.0.3
openai-whisper==20231117
librosa==0.10.2.post1
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.4.0
python-magic==0.4.27
numpy==1.26.4
gunicorn==23.0.0
python-dotenv==1.0.1
werkzeug==3.0.3
torch==2.3.1
torchaudio==2.3.1
soundfile